% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_name_lookup.R
\name{build_name_lookup}
\alias{build_name_lookup}
\title{Build a name-lookup .vtr from a backbone .vtr}
\usage{
build_name_lookup(bb_path, out_path, verbose = TRUE)
}
\arguments{
\item{bb_path}{Character. Backbone .vtr path.}

\item{out_path}{Character. Lookup .vtr destination.}

\item{verbose}{Logical.}
}
\value{
\code{out_path} (invisibly).
}
\description{
Build a name-lookup .vtr from a backbone .vtr
}
