% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-ncbi.R
\name{build_ncbi}
\alias{build_ncbi}
\title{Build the NCBI backbone .vtr}
\usage{
build_ncbi(output_dir = "output/ncbi", version = NULL, verbose = TRUE)
}
\arguments{
\item{output_dir}{Character.}

\item{version}{Character or NULL.}

\item{verbose}{Logical.}
}
\value{
Path to the .vtr file (invisibly).
}
\description{
Build the NCBI backbone .vtr
}
