% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-reptiledb.R
\name{build_reptiledb}
\alias{build_reptiledb}
\title{Build the Reptile Database backbone .vtr}
\usage{
build_reptiledb(
  output_dir = "output/reptiledb",
  version = NULL,
  verbose = TRUE
)
}
\arguments{
\item{output_dir}{Character.}

\item{version}{Character or NULL.}

\item{verbose}{Logical.}
}
\value{
Path to the .vtr file (invisibly).
}
\description{
Build the Reptile Database backbone .vtr
}
