% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{count_vtr_rows}
\alias{count_vtr_rows}
\title{Count rows in a .vtr file}
\usage{
count_vtr_rows(vtr_path)
}
\arguments{
\item{vtr_path}{Character.}
}
\value{
Integer.
}
\description{
Count rows in a .vtr file
}
