% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-avilist.R
\name{download_avilist}
\alias{download_avilist}
\title{Download the AviList extended checklist workbook}
\usage{
download_avilist(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Path to the downloaded \code{.xlsx}.
}
\description{
Download the AviList extended checklist workbook
}
