% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-col.R
\name{download_col}
\alias{download_col}
\title{Download and extract the COL DwC-A}
\usage{
download_col(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Path to the extracted COL directory.
}
\description{
Download and extract the COL DwC-A
}
