% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-fungorum.R
\name{download_fungorum}
\alias{download_fungorum}
\title{Download and extract the Species Fungorum Plus ColDP archive}
\usage{
download_fungorum(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Path to the directory containing the extracted ColDP files.
}
\description{
Download and extract the Species Fungorum Plus ColDP archive
}
