% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-gbif.R
\name{download_gbif}
\alias{download_gbif}
\title{Download the GBIF backbone simple.txt.gz}
\usage{
download_gbif(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Path to the .gz file.
}
\description{
Download the GBIF backbone simple.txt.gz
}
