% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-itis.R
\name{download_itis}
\alias{download_itis}
\title{Download and extract the ITIS SQLite database}
\usage{
download_itis(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Path to the extracted SQLite database file.
}
\description{
Download and extract the ITIS SQLite database
}
