% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-lcvp.R
\name{download_lcvp}
\alias{download_lcvp}
\title{Download the LCVP \code{tab_lcvp} data file}
\usage{
download_lcvp(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Path to the downloaded \code{tab_lcvp.rda}.
}
\description{
Download the LCVP \code{tab_lcvp} data file
}
