% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-lpsn.R
\name{download_lpsn}
\alias{download_lpsn}
\title{Download the LPSN ColDP archive from ChecklistBank}
\usage{
download_lpsn(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Path to the directory holding the extracted \code{NameUsage.tsv}.
}
\description{
Download the LPSN ColDP archive from ChecklistBank
}
