% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-mdd.R
\name{download_mdd}
\alias{download_mdd}
\title{Download and unpack the Mammal Diversity Database archive}
\usage{
download_mdd(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Path to the directory holding the extracted CSVs.
}
\description{
Download and unpack the Mammal Diversity Database archive
}
