% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-ncbi.R
\name{download_ncbi}
\alias{download_ncbi}
\title{Download and extract NCBI taxdump}
\usage{
download_ncbi(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Path to the extraction directory.
}
\description{
Download and extract NCBI taxdump
}
