% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-ott.R
\name{download_ott}
\alias{download_ott}
\title{Download and extract OTT taxonomy}
\usage{
download_ott(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Path to the extraction directory containing taxonomy.tsv.
}
\description{
Download and extract OTT taxonomy
}
