% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-reptiledb.R
\name{download_reptiledb}
\alias{download_reptiledb}
\title{Download the Reptile Database bulk files}
\usage{
download_reptiledb(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Named list with paths: \code{taxa}, \code{synonyms}, \code{checklist}.
}
\description{
Download the Reptile Database bulk files
}
