% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-wcvp.R
\name{download_wcvp}
\alias{download_wcvp}
\title{Download and extract the WCVP names file}
\usage{
download_wcvp(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Path to the extracted \code{wcvp_names.csv}.
}
\description{
Download and extract the WCVP names file
}
