% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-wgsrpd.R
\name{download_wgsrpd}
\alias{download_wgsrpd}
\title{Download the WGSRPD Level 3 GeoJSON}
\usage{
download_wgsrpd(dest = tempfile(fileext = ".geojson"), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Output path for the downloaded GeoJSON.}

\item{verbose}{Logical.}
}
\value{
The path to the downloaded GeoJSON (invisibly).
}
\description{
Download the WGSRPD Level 3 GeoJSON
}
