% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-worms.R
\name{download_worms}
\alias{download_worms}
\title{Download and extract WoRMS DwC-A}
\usage{
download_worms(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Path to the extracted DwC-A directory.
}
\description{
Download and extract WoRMS DwC-A
}
