% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_backend.R
\name{list_backends}
\alias{list_backends}
\title{List available backend builders}
\usage{
list_backends()
}
\value{
Character vector of backend identifiers.
}
\description{
List available backend builders
}
