% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_enrichment.R
\name{list_enrichments}
\alias{list_enrichments}
\title{List available enrichment builders}
\usage{
list_enrichments()
}
\value{
Character vector of enrichment identifiers.
}
\description{
List available enrichment builders
}
