% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment-parsers-traits.R
\name{parse_anage}
\alias{parse_anage}
\title{Parse AnAge longevity and life-history traits (TSV from ZIP)}
\usage{
parse_anage(path)
}
\arguments{
\item{path}{Character. Path to anage_*.txt.}
}
\value{
data.frame with canonical_name + AnAge trait columns.
}
\description{
Parse AnAge longevity and life-history traits (TSV from ZIP)
}
