% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment-parsers.R
\name{parse_elton_traits}
\alias{parse_elton_traits}
\title{Parse EltonTraits 1.0 birds + mammals TSVs}
\usage{
parse_elton_traits(birds_path, mammals_path)
}
\arguments{
\item{birds_path}{Character. Path to BirdFuncDat.txt.}

\item{mammals_path}{Character. Path to MamFuncDat.txt.}
}
\value{
data.frame with canonical_name + diet/foraging/body mass columns.
}
\description{
Parse EltonTraits 1.0 birds + mammals TSVs
}
