% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment-parsers-names.R
\name{parse_ncbi_common_names}
\alias{parse_ncbi_common_names}
\title{Parse NCBI common names from names.dmp}
\usage{
parse_ncbi_common_names(dir_path)
}
\arguments{
\item{dir_path}{Character. Directory containing names.dmp.}
}
\value{
data.frame with canonical_name + lang (NA) + common_name.
}
\description{
Parse NCBI common names from names.dmp
}
