% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment-parsers-names.R
\name{parse_ott_common_names}
\alias{parse_ott_common_names}
\title{Parse OTT common names from synonyms.tsv + taxonomy.tsv}
\usage{
parse_ott_common_names(dir_path)
}
\arguments{
\item{dir_path}{Character. Directory containing synonyms.tsv + taxonomy.tsv.}
}
\value{
data.frame with canonical_name + lang (NA) + common_name.
}
\description{
Parse OTT common names from synonyms.tsv + taxonomy.tsv
}
