% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment-parsers.R
\name{parse_woodiness}
\alias{parse_woodiness}
\title{Parse Zanne et al. 2014 woodiness CSV}
\usage{
parse_woodiness(path)
}
\arguments{
\item{path}{Character. Path to GlobalWoodinessDatabase.csv.}
}
\value{
data.frame with canonical_name + woodiness.
}
\description{
Parse Zanne et al. 2014 woodiness CSV
}
