% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-avilist.R
\name{read_avilist}
\alias{read_avilist}
\title{Read and normalize the AviList checklist}
\usage{
read_avilist(xlsx_path, verbose = TRUE)
}
\arguments{
\item{xlsx_path}{Character. Path to the AviList extended workbook.}

\item{verbose}{Logical.}
}
\value{
A normalized data.frame ready for \code{\link[=precompute_backbone]{precompute_backbone()}}.
}
\description{
Read and normalize the AviList checklist
}
