% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-col.R
\name{read_col}
\alias{read_col}
\title{Read and normalize the COL Taxon.tsv}
\usage{
read_col(col_dir, verbose = TRUE)
}
\arguments{
\item{col_dir}{Character. Path to the extracted COL directory.}

\item{verbose}{Logical.}
}
\value{
A normalized data.frame.
}
\description{
Read and normalize the COL Taxon.tsv
}
