% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-ncbi.R
\name{read_dmp}
\alias{read_dmp}
\title{Read a pipe-delimited NCBI .dmp file}
\usage{
read_dmp(path, col_names)
}
\arguments{
\item{path}{Character. Path to the .dmp file.}

\item{col_names}{Character vector of column names.}
}
\value{
A data.frame.
}
\description{
Read a pipe-delimited NCBI .dmp file
}
