% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-fungorum.R
\name{read_fungorum}
\alias{read_fungorum}
\title{Read and normalize the Species Fungorum Plus ColDP files}
\usage{
read_fungorum(fungorum_dir, verbose = TRUE)
}
\arguments{
\item{fungorum_dir}{Character. Directory containing Taxon.tsv, Name.tsv,
Synonym.tsv.}

\item{verbose}{Logical.}
}
\value{
A normalized data.frame.
}
\description{
Read and normalize the Species Fungorum Plus ColDP files
}
