% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-itis.R
\name{read_itis}
\alias{read_itis}
\title{Read and normalize the ITIS SQLite database}
\usage{
read_itis(sqlite_path, verbose = TRUE)
}
\arguments{
\item{sqlite_path}{Character. Path to the ITIS .sqlite file.}

\item{verbose}{Logical.}
}
\value{
A normalized data.frame ready for \code{\link[=precompute_backbone]{precompute_backbone()}}.
}
\description{
Read and normalize the ITIS SQLite database
}
