% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-lcvp.R
\name{read_lcvp}
\alias{read_lcvp}
\title{Read and normalize the LCVP reference table}
\usage{
read_lcvp(rda_path, verbose = TRUE)
}
\arguments{
\item{rda_path}{Character. Path to \code{tab_lcvp.rda}.}

\item{verbose}{Logical.}
}
\value{
A normalized data.frame ready for \code{\link[=precompute_backbone]{precompute_backbone()}}.
}
\description{
Read and normalize the LCVP reference table
}
