% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-lpsn.R
\name{read_lpsn}
\alias{read_lpsn}
\title{Read and normalize the LPSN checklist}
\usage{
read_lpsn(lpsn_dir, verbose = TRUE)
}
\arguments{
\item{lpsn_dir}{Character. Directory holding \code{NameUsage.tsv}.}

\item{verbose}{Logical.}
}
\value{
A normalized data.frame ready for \code{\link[=precompute_backbone]{precompute_backbone()}}.
}
\description{
Read and normalize the LPSN checklist
}
