% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-ncbi.R
\name{read_ncbi}
\alias{read_ncbi}
\title{Read and normalize the NCBI taxonomy dump}
\usage{
read_ncbi(dump_dir, verbose = TRUE)
}
\arguments{
\item{dump_dir}{Character. Path to the extracted taxdump directory.}

\item{verbose}{Logical.}
}
\value{
A normalized data.frame ready for \code{\link[=precompute_backbone]{precompute_backbone()}}.
}
\description{
Read and normalize the NCBI taxonomy dump
}
