% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-ott.R
\name{read_ott}
\alias{read_ott}
\title{Read and normalize the OTT taxonomy}
\usage{
read_ott(ott_dir, verbose = TRUE)
}
\arguments{
\item{ott_dir}{Character. Path to the extracted OTT directory.}

\item{verbose}{Logical.}
}
\value{
A normalized data.frame.
}
\description{
Read and normalize the OTT taxonomy
}
