% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-reptiledb.R
\name{read_reptiledb}
\alias{read_reptiledb}
\title{Read and normalize the Reptile Database}
\usage{
read_reptiledb(paths, verbose = TRUE)
}
\arguments{
\item{paths}{Named list from \code{\link[=download_reptiledb]{download_reptiledb()}}.}

\item{verbose}{Logical.}
}
\value{
A normalized backbone data.frame.
}
\description{
Read and normalize the Reptile Database
}
