% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-wcvp.R
\name{read_wcvp}
\alias{read_wcvp}
\title{Read and normalize the WCVP names file}
\usage{
read_wcvp(names_path, verbose = TRUE)
}
\arguments{
\item{names_path}{Character. Path to \code{wcvp_names.csv}.}

\item{verbose}{Logical.}
}
\value{
A normalized data.frame ready for \code{\link[=precompute_backbone]{precompute_backbone()}}.
}
\description{
Read and normalize the WCVP names file
}
