% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-wfo.R
\name{read_wfo}
\alias{read_wfo}
\title{Read and normalize the WFO classification file}
\usage{
read_wfo(txt_path, verbose = TRUE)
}
\arguments{
\item{txt_path}{Character. Path to the WFO classification.txt file.}

\item{verbose}{Logical.}
}
\value{
A normalized data.frame ready for \code{\link[=precompute_backbone]{precompute_backbone()}}.
}
\description{
Read and normalize the WFO classification file
}
