% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-worms.R
\name{read_worms}
\alias{read_worms}
\title{Read and normalize the WoRMS taxonomy}
\usage{
read_worms(worms_dir, verbose = TRUE)
}
\arguments{
\item{worms_dir}{Character. Path to the extracted DwC-A directory.}

\item{verbose}{Logical.}
}
\value{
A normalized data.frame.
}
\description{
Read and normalize the WoRMS taxonomy
}
