% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_versions.R
\name{check_dryad_version}
\alias{check_dryad_version}
\title{Check a Dryad dataset for the latest version}
\usage{
check_dryad_version(source_url)
}
\arguments{
\item{source_url}{Character. Dryad download URL containing a DOI.}
}
\value{
Named list with \code{version} and \code{url}.
}
\description{
Check a Dryad dataset for the latest version
}
