% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_versions.R
\name{check_figshare_version}
\alias{check_figshare_version}
\title{Check a Figshare article for the latest version}
\usage{
check_figshare_version(source_url)
}
\arguments{
\item{source_url}{Character. Figshare download URL.}
}
\value{
Named list with \code{version} and \code{url}.
}
\description{
Check a Figshare article for the latest version
}
