% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_versions.R
\name{check_gbif_version}
\alias{check_gbif_version}
\title{Check a GBIF hosted dataset for the Last-Modified date}
\usage{
check_gbif_version(source_url)
}
\arguments{
\item{source_url}{Character. GBIF hosted dataset URL.}
}
\value{
Named list with \code{version} (date string) and \code{url}.
}
\description{
Check a GBIF hosted dataset for the Last-Modified date
}
