% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_versions.R
\name{check_wcvp_version}
\alias{check_wcvp_version}
\title{Check Kew WCVP for the latest version}
\usage{
check_wcvp_version(source_url)
}
\arguments{
\item{source_url}{Character. WCVP download URL.}
}
\value{
Named list with \code{version} and \code{url}.
}
\description{
WCVP has no dedicated API, so we fall back to a HEAD request via
\code{\link[=check_gbif_version]{check_gbif_version()}}.
}
