% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_versions.R
\name{check_zenodo_version}
\alias{check_zenodo_version}
\title{Check a Zenodo record for the latest version}
\usage{
check_zenodo_version(source_url)
}
\arguments{
\item{source_url}{Character. Zenodo download URL.}
}
\value{
Named list with \code{version} (publication date) and \code{url}.
}
\description{
Check a Zenodo record for the latest version
}
