% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-euromed.R
\name{download_euromed}
\alias{download_euromed}
\title{Download and extract the Euro+Med CSV}
\usage{
download_euromed(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Path to the extracted CSV.
}
\description{
Download and extract the Euro+Med CSV
}
