% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-wfo.R
\name{download_wfo}
\alias{download_wfo}
\title{Download and extract the WFO classification file}
\usage{
download_wfo(dest = tempdir(), verbose = TRUE)
}
\arguments{
\item{dest}{Character. Destination directory.}

\item{verbose}{Logical.}
}
\value{
Path to the extracted classification file.
}
\description{
Download and extract the WFO classification file
}
