% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-gbif.R
\name{read_gbif}
\alias{read_gbif}
\title{Read and normalize the GBIF backbone}
\usage{
read_gbif(gz_path, verbose = TRUE)
}
\arguments{
\item{gz_path}{Character. Path to simple.txt.gz.}

\item{verbose}{Logical.}
}
\value{
A normalized data.frame ready for \code{\link[=precompute_backbone]{precompute_backbone()}}.
}
\description{
Read and normalize the GBIF backbone
}
